Moved code from common.py to mpienv/__init__.py
